+++ app/models/student_lesson_summary.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+# rubocop:disable Rails/ApplicationRecord
+class StudentLessonSummary < ActiveRecord::Base
+  def readonly?
+    true
+  end
+end
+# rubocop:enable Rails/ApplicationRecord